webapi/webapi-contactsmanager-w3c-tests/contactsmanager/ContactsManager_save_Contact_fullset.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xu, Kang <kangx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>ContactsManager Test: ContactsManager_save_Contact_fullset</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="http://www.w3.org/2012/sysapps/contacts-manager-api/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>

var t = async_test("Check if save method of ContactsManager can work properly with contact full set", {timeout: 2000});
t.step(function () {
  contacts.clear().then(function () {/* Do Nothing */}, function () {/* Do Nothing */});
  var url = new ContactField({
    types: ['work'],
    preferred: true,
    value: 'http://www.intel.com/'
  });
  var emailAddr = new ContactField({
    types: ['work'],
    preferred: true,
    value: 'tester@sample.com'
  });
  var address = new ContactAddress({
    types: ['work'],
    preferred: true,
    countryName: 'china'
  });
  var mobilePhone = new ContactTelField({
    types: ['home'],
    preferred: true,
    value: '+34698765432'
  });
  var instantAddr = new ContactField({
    types: ['home'],
    preferred: true,
    value: 'msn:tester@sample.com'
  });
  var data = new Date();
  var dataTime = data.getTime();
  var contactName = new ContactName({
    givenNames: ['John'],
    familyNames: ['Doe']
  });
  var contact = new Contact({
    name: contactName,
    phoneNumbers: [mobilePhone],
    urls: [url],
    emails: [emailAddr],
    photos: ["test"],
    categories: ['swimmer,biker'],
    addresses: [address],
    organizations: ["Bubba Gump Shrimp Co."],
    jobTitles: ["Shrimp Man"],
    birthday: data,
    notes: ["I am proficient in Tiger-Crane Style"],
    impp: [instantAddr],
    anniversary: data,
    gender: 'male'
  });
  contacts.save(contact).then(
    function (info) {
      t.step(function () {
        assert_equals(info.name.familyNames[0], "Doe", "The Contact.name.familyNames[0]");
        assert_equals(info.phoneNumbers[0].value, "+34698765432", "The Contact.phoneNumbers[0].value");
        assert_equals(info.urls[0].value, "http://www.intel.com/", "The Contact.urls[0].value");
        assert_equals(info.emails[0].value, "tester@sample.com", "The Contact.emails[0].value");
        assert_equals(info.photos[0], "test", "The Contact.photos[0]");
        assert_equals(info.categories[0], "swimmer,biker", "The Contact.categories[0]");
        assert_equals(info.addresses[0].countryName, "china", "The Contact.addresses[0].countryName");
        assert_equals(info.organizations[0], "Bubba Gump Shrimp Co.", "The Contact.organizations[0]");
        assert_equals(info.jobTitles[0], "Shrimp Man", "The Contact.jobTitles[0]");
        assert_equals(info.birthday.getTime(), dataTime, "The Contact.birthday");
        assert_equals(info.notes[0], "I am proficient in Tiger-Crane Style", "The Contact.notes[0]");
        assert_equals(info.impp[0].value, "msn:tester@sample.com", "The Contact.impp[0].value");
        assert_equals(info.anniversary.getTime(), dataTime, "The Contact.anniversary");
        assert_equals(info.gender, "male", "The Contact.gender attribute");
      });
      t.done();
    },
    function (error) {
      t.step(function () {
        assert_unreached(error.message);
      });
      t.done();
    }
  );
});

</script>
